+++ src/main/java/domain/member/Money.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain.member;
 
 import java.util.Objects;
