+++ src/test/java/domain/card/MoneyTest.java
@@ -1,5 +1,6 @@
 package domain.card;
 
+import domain.member.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
